requirements.txt
pyqt5
pyserial
pyaudio
numpy
python-osc
adafruit-circuitpython-dht
RPi.GPIO